Delete reference to FAQ pages on court website
this bit about modifying child support in changing custody and changing divorce orders is only for responding to motions to modify, it is lifted almost word for word from child support interview but the sentence is only for users filing a motion for child support not for responding. So deleted for now.